Move point helpers.ts into view/points/util.
From view/pointsmanagement/util.
