+++ client/src/view/points/util/helper.ts
